# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.PaginatedStatementResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      statements:
        items:
          $ref: '#/definitions/models.StatementResponse'
        type: array
      total:
        type: integer
    type: object
...
  models.StatementPreview:
    properties:
      headers:
        items:
          type: string
        type: array
      rows:
        items:
          items:
            type: string
          type: array
        type: array
    type: object
  models.StatementResponse:
    properties:
      account_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: integer
      file_type:
        type: string
      id:
        type: integer
      message:
        type: string
      original_filename:
        type: string
      status:
        $ref: '#/definitions/models.StatementStatus'
    type: object
  models.StatementStatus:
    enum:
    - pending
    - processing
    - done
    - error
    type: string
    x-enum-varnames:
    - StatementStatusPending
    - StatementStatusProcessing
    - StatementStatusDone
    - StatementStatusError
...
host: localhost:5000
info:
  contact: {}
  description: Enter your email as username and password to authenticate
  title: Expense Tracker API
  version: "1.0"
paths:
...
  /statement:
    get:
      description: Get all uploaded statements for the authenticated user with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 15
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of statements
          schema:
            $ref: '#/definitions/models.PaginatedStatementResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BasicAuth: []
      summary: List statements
      tags:
      - statements
    post:
      consumes:
      - multipart/form-data
      description: Upload and process a bank statement file for transaction parsing
      parameters:
      - description: Account ID
        in: formData
        name: account_id
        required: true
        type: integer
      - description: Bank type
        enum:
        - investment
        - axis
        - sbi
        - hdfc
        - icici
        - others
        in: formData
        name: bank_type
        required: true
        type: string
      - description: Additional metadata
        in: formData
        name: metadata
        type: string
      - description: Statement file (CSV, PDF, etc.)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Statement uploaded and processing started
          schema:
            $ref: '#/definitions/models.StatementResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BasicAuth: []
      summary: Upload bank statement
      tags:
      - statements
  /statement/{id}:
    get:
      description: Get the processing status and details of a specific statement
      parameters:
      - description: Statement ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Statement status and details
          schema:
            $ref: '#/definitions/models.StatementResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Statement not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BasicAuth: []
      summary: Get statement status
      tags:
      - statements
  /statement/preview:
    post:
      consumes:
      - multipart/form-data
      description: Preview the contents of a bank statement file before processing
      parameters:
      - description: Number of rows to skip
        in: formData
        name: skip_rows
        type: integer
      - description: Number of rows to preview
        in: formData
        name: row_size
        type: integer
      - description: Statement file (CSV, PDF, etc.)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Statement preview
          schema:
            $ref: '#/definitions/models.StatementPreview'
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BasicAuth: []
      summary: Preview bank statement
      tags:
      - statements
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"